Like a phoenix the RS Latch will be reborn.
I completly threw away the old rs latch logic, because it was broken and
so I made a new RS Latch that works great and true to life. South, West,
and East are inputs for the Q, R, S. North is an output for the new
state of Q. Also updated the textures to match the changes.